+++ scholarshippro/scholarships/urls.py
@@ -5,7 +5,6 @@ from . import views
 urlpatterns = [ 
     path('scholarships/', views.ScholarshipsList.as_view()), 
     path('scholarships/<int:pk>/', views.ScholarshipDetail.as_view()),
-    path('scholarships/favorite/<int:pk>/', views.FavoriteDetail.as_view()), 
 ]
 
 urlpatterns = format_suffix_patterns(urlpatterns)
