+++ scholarshippro/scholarships/views.py
@@ -98,38 +98,6 @@ class ScholarshipDetail(APIView):
             status=status.HTTP_401_UNAUTHORIZED
         )
 
-class FavoriteDetail(APIView):
-
-    def get_object(self, pk):
-        try:
-            scholarship = Scholarships.objects.get(pk=pk)
-            return scholarship
-        except Scholarships.DoesNotExist:
-            raise Http404
-
-    def post(self, request, pk):
-        scholarship = self.get_object(pk)
-        
-        if request.user not in scholarship.favorites.all():
-            scholarship.favorites.add(request.user)
-            return Response(
-            status=status.HTTP_201_CREATED
-            )
-        return Response(
-            status=status.HTTP_400_BAD_REQUEST
-        )
-
-    def delete(self, request, pk):
-        scholarship = Scholarships.objects.get(pk=pk)
-        
-        if request.user in scholarship.favorites.all():
-            scholarship.favorites.remove(request.user)
-            return Response(
-            status=status.HTTP_204_NO_CONTENT
-            )
-        return Response(
-            status=status.HTTP_400_BAD_REQUEST
-        )
