+++ scholarshippro/users/urls.py
@@ -4,7 +4,8 @@ from . import views
 
 urlpatterns = [
     path('users/', views.CustomUserList.as_view()),
-    path('users/<int:pk>/', views.CustomUserDetail.as_view()),   
+    path('users/<int:pk>/', views.CustomUserDetail.as_view()), 
+    path('scholarships/favorite/<int:pk>/', views.FavoriteDetail.as_view()),   
 ]
 
 urlpatterns = format_suffix_patterns(urlpatterns)
