+++ scholarshippro/users/views.py
@@ -8,6 +8,7 @@ from rest_framework.response import Response
 from rest_framework import status
 from .models import CustomUser
 from .serializers import CustomUserSerializer
+from scholarships.models import Scholarships 
 
 class CustomUserList(APIView):
 
@@ -66,4 +67,37 @@ class CustomUserDetail(APIView):
             return Response(status=status.HTTP_204_NO_CONTENT)
         return Response(
             status=status.HTTP_401_UNAUTHORIZED
+        )
+
+class FavoriteDetail(APIView):
+
+    def get_object(self, pk):
+        try:
+            scholarship = Scholarships.objects.get(pk=pk)
+            return scholarship
+        except Scholarships.DoesNotExist:
+            raise Http404
+
+    def post(self, request, pk):
+        scholarship = self.get_object(pk)
+        
+        if scholarship not in request.user:
+            request.user.add(scholarship)
+            return Response(
+            status=status.HTTP_201_CREATED
+            )
+        return Response(
+            status=status.HTTP_400_BAD_REQUEST
+        )
+
+    def delete(self, request, pk):
+        scholarship = Scholarships.objects.get(pk=pk)
+        
+        if scholarship in request.user:
+            request.user.remove(scholarship)
+            return Response(
+            status=status.HTTP_204_NO_CONTENT
+            )
+        return Response(
+            status=status.HTTP_400_BAD_REQUEST
         )
